\name{arcene}
\docType{data}
\alias{arcene}
\title{Normalized version of the Arcene data set}
\description{
  This is a normalized version of the publically available Arcene mass-spectrometry data set which contains 9961 variables and compares the mass spectrometry data of cancer (1, n=88) versus normal subjects (2, n=112). Noteworthy, the Arcene dataset was one of the datasets used in Neural Information Processing Systems (NIPS) 2003 Feature Selection Challenge and contains 3000 'probes' - added distractor features (of random orders and patterns) that are having no predictive power (pure noise).
}
\usage{arcene}
\format{data.frame containing 200 observations of 9961 variables.}
\references{
  \url{https://archive.ics.uci.edu/ml/datasets/Arcene}
}
\keyword{datasets}
