\name{Clust.exampleE}
\alias{Clust.exampleE}
\docType{data}
\title{ The example data of the mean gene expression levels}
\description{
  This data is an artificial example of the mean gene expression levels.
}
\usage{data(Clust.exampleE)}
\format{
  A 700x20 matrix including 700 genes and 20 chips. Every 100 genes belong to one cluster from the first gene. 
  There are 7 clusters.
}
\source{
Liu, X. Lin, K., Andersen, B. Rattray, M. (2007) Including probe-level uncertainty in model-based gene expression clustering, BMC Bioinformatics, 8(98).
}
\seealso{\code{\link{Clust.exampleStd}}}
\keyword{datasets}
