\name{Clustii.exampleStd}
\alias{Clustii.exampleStd}
\docType{data}
\title{ The example data of the standard deviation for gene expression levels }
\description{
  This data is an artificial example of the standard deviation for gene exapression levels generated 
  by package \code{\link[puma]{mmgmos}}.
}
\usage{data(Clustii.exampleStd)}
\format{
  A 600x80 matrix including 600 genes and 20 conditions. Each condition has 4 replicates. Every 100 genes belong to one cluster from the first gene. 
  There are 6 clusters.
}
\source{
Liu,X. and Rattray,M. (2009) 
Including probe-level measurement error in robust mixture clustering of replicated microarray gene expression, 
technical report available upon request.
}
\seealso{\code{\link[puma]{Clustii.exampleE}}}
\keyword{datasets}
