\name{coffee}
\docType{data}
\alias{coffee}
\title{Normalized version of the Coffee data set}
\description{
  This is a normalized verion of the Coffee data set which contains nuclear magnetic resonance (NMR) profiles of Colombian (n=15) vs non-Colombian (n=19) coffee bean samples using 1175 normalized spectral bins.
}
\usage{coffee}
\format{data.frame containing 34 observations of 1175 variables.}
\references{
  \url{https://doi.org/10.1155/2016/8564584}
}
\keyword{datasets}

