\name{diabetes}
\docType{data}
\alias{diabetes}
\title{Normalized version of the QMDiab data set}
\description{
  This is a normalized verion of The Qatar Metabolomics Study on Diabetes (QMDiab) publicly available data set containing untargeted plasma metabolomics (LC/MS+, LC/MS-, and GC/MS) from 179 diabetic patients and 177 non-diabetic normal controls.
}
\usage{diabetes}
\format{data.frame containing 356 observations of 322 variables.}
\references{
  \url{https://doi.org/10.6084/m9.figshare.5904022.v5}
}
\keyword{datasets}
