\name{exampleE}
\alias{exampleE}
\docType{data}
\title{ The example data of the mean gene expression levels}
\description{
  This data is an artificial example of the mean gene expression levels from 
  golden spike-in data set in Choe et al. (2005).
}
\usage{data(exampleE)}
\format{
  A 200x6 matrix including 200 genes and 6 chips. The first 3 chips are replicates for 
  C condition and the last 3 chips are replicates for S conditon.
}
\source{
  Choe,S.E., Boutros,M., Michelson,A.M., Church,G.M., Halfon,M.S.: 
  Preferred analysis methods for Affymetrix GeneChips revealed by a wholly defined control dataset. 
  Genome Biology, 6 (2005) R16.
}
\seealso{\code{\link{exampleStd}}}
\keyword{datasets}
