\name{license.puma}
\alias{license.puma}
\title{Print puma license}
\description{
  This function prints the license under which puma is made available.
}
\usage{
license.puma()
}
%\arguments{
%}
%\details{
%}
\value{
Null.

}
\author{
Richard Pearson (based on the license.cosmo function from the cosmo package)
}
%\seealso{
%}
\examples{
license.puma()
}
\keyword{misc}
