\name{mice}
\docType{data}
\alias{mice}
\title{Normalized version of the Mice data set}
\description{
  This is a normalized version of the Mice data set which compares the difference in 28 proteins between control mice (c-SC-s: control mice, not stimulated to learn, injected with saline. n=135) and mice with trisomy 21 i.e. Down syndrome (t-CS-m: trisomy mice, stimulated to learn, injected with memantine. n=135).
}
\usage{mice}
\format{data.frame containing 270 observations of 28 variables.}
\references{
  \url{https://doi.org/10.6084/m9.figshare.1421985.v1}
}
\keyword{datasets}

